Time: 7 ms (84.20%) | Memory: 17.6 MB (42.71%) - LeetSync

+++ 1556-make-two-arrays-equal-by-reversing-subarrays/make-two-arrays-equal-by-reversing-subarrays.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     bool canBeEqual(vector<int>& target, vector<int>& arr) {
-        int count=0;
-        int n=target.size();
+        long count=0;
+        long n=target.size();
         sort(target.begin(),target.end());
         sort(arr.begin(),arr.end());
         for(int i=0;i<n;i++){
